Refactor cpp protocol generation to its own file in wayland-rs
