+++ src/server/frontend_wayland/wayland_rs/build_script/ffi_generation.rs
@@ -11,7 +11,7 @@ use crate::cpp_builder::{sanitize_identifier, CppBuilder};
 use crate::helpers::format_wayland_interface_to_rust_extension_struct;
 
 use super::protocol_middleware_generation::wayland_arg_to_ffi_rust_str;
-use super::{InterfaceItem, WaylandEvent, WaylandInterface, WaylandProtocol};
+use crate::protocol_parser::{InterfaceItem, WaylandEvent, WaylandInterface, WaylandProtocol};
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
